Fix character info visibility logic
The following code is buggy:

  showOverview = showOverview == 1 and characterViewable == true

showOverview converts from integer to boolean and on the next
iteration the variable becomes false, because true == 1 is false.

The bug is can be seen in multiplayer games, where updateAttrTable
gets called repeatedly without resetting showX variables.

Also update the record variable on each updateAttrTable just in case.