Improve performance of spellbooks
Remove {priestScreen,mageScreen}:SetSpellLevel.

The function takes 25ms to run and on every menu file reload the time increases
by a few ms. After many reloads it adds up and becomes noticeable. I don't know
what it does, at the very least it calls refreshMageBook or refreshPriestBook,
which is faster to do from Lua code, takes only about 1 ms.

There is no priestScreen:SetSpellLevel anywhere in the unmodded UI.menu file
and everything seems to work just fine.

As a side note, everytime a mage opens the arcane spellbook refreshMageBook is
called 4 times and if a priest opens it there will also be additional 4 calls
to refreshPriestBook. This seems to be hardcoded.